Refactored code:Purchase2Items.feature,Add Hooks,Add DriverFactory,Remove most Thread.sleep,Created a BaseStepDefs,Created seperate Step definition files for the features
