+++ src/test/java/helper/DriverFactory.java
@@ -7,32 +7,25 @@ import org.openqa.selenium.chrome.ChromeDriver;
 
 import java.util.concurrent.TimeUnit;
 
-public class BasePage {
+public class DriverFactory {
 
     public static WebDriver driver;
 
-    public  String baseURL = "http://automationpractice.com/";
-
-
-    public void openChrome(String url) throws Exception {
+    public void openChrome()  {
         System.setProperty("webdriver.chrome.driver",
                 System.getProperty("user.dir") + "//src//main//resources//drivers/chromedriver");
         driver = new ChromeDriver();
         driver.manage().window().maximize();
-        driver.manage().timeouts().implicitlyWait(15, TimeUnit.SECONDS);
-        driver.get(url);
+        driver.get("http://automationpractice.com/");
+        driver.manage().timeouts().implicitlyWait(100, TimeUnit.SECONDS);
+
     }
+
     public void scrollDown (){
         JavascriptExecutor js = (JavascriptExecutor) driver;
         js.executeScript("window.scrollBy(0,500)");
 
     }
 
-    @After
-    public void teardown() {
-        driver.quit();
-
-
-    }
 
 }
