+++ src/test/java/pageObject/AddressPage.java
@@ -1,13 +1,13 @@
 package pageObject;
 
-import helper.BasePage;
-import org.openqa.selenium.By;
+import helper.DriverFactory;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
-public class AddressPage extends BasePage {
+public class AddressPage {
+    WebDriver driver;
 
     @FindBy(css = "button.button:nth-child(4) > span:nth-child(1)")
     private WebElement proceedToCheckOutButton;
